refactor: fix type(used days: int -> double)

+++ src/main/java/com/croquiscom/recruit/config/MemberSettingConfig.java
@@ -2,6 +2,6 @@ package com.croquiscom.recruit.config;
 
 public abstract class MemberSettingConfig {
 
-    public static final Integer TOTAL_VACATION_DAYS = 15;
+    public static final Double TOTAL_VACATION_DAYS = 15D;
 
 }
